src/sailing_club/auth/models/user.py
from sqlalchemy.orm import declarative_base
from sailing_club.app import db

Base = declarative_base()

class User(db.Model):
    __tablename__ = 'user'

    id = db.Column(db.Integer, primary_key=True, autoincrement=True)
    username = db.Column(db.String(), unique=True, nullable=False)
    password = db.Column(db.String(), nullable=False)
    created_at = db.Column(db.Date, nullable=False)