pyproject.toml
[project]
name = "sailing-club"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "benjamin-rosenthal", email = "brosen8690@gmail.com" }
]
dependencies = [
    "pip>=24.0",
    "flask>=3.0.3",
    "flask-jwt-extended>=4.6.0",
    "python-dotenv>=1.0.1",
    "setuptools>=69.5.1",
    "flask-migrate>=4.0.7",
    "flask-sqlalchemy>=3.1.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/sailing_club"]
